Added mapping the users to the database
