+++ package.json
@@ -10,6 +10,7 @@
     "postinstall": "prisma generate"
   },
   "dependencies": {
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^5.10.2",
     "@vercel/speed-insights": "^1.0.10",
     "next": "14.1.1",
