+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,16 +1,6 @@
-import { NextAuthOptions } from "next-auth";
 import NextAuth from "next-auth/next";
-import GithubProvider from "next-auth/providers/github"
-
-const authOptions: NextAuthOptions = {
-    providers: [
-        GithubProvider({
-            clientId: process.env.GITHUB_ID!,
-            clientSecret: process.env.GITHUB_SECRET!,
-        }),
-    ],
-}
+import authOptions from "@/../lib/auth";
 
 const handler = NextAuth(authOptions);
 
-export {handler as GET, handler as POST, authOptions}
+export {handler as GET, handler as POST};
